docker/config/requirements.txt
customtkinter
pyyaml